Refactor context into separate actions/reducers files
