+++ src/constants/index.tsx
@@ -1,3 +1,4 @@
 // Date Format Constants
 export const MMMD = "MMM d";
 export const MMMMdyyyy = "MMMM d, yyyy";
+export const STATE_PRISON = "State Prison";
